+++ index.js
@@ -32,6 +32,14 @@ export default {
         ignoreAtRules: ['value'],
       },
     ],
+    'at-rule-empty-line-before': [
+      'always',
+      {
+        except: ['blockless-after-same-name-blockless', 'first-nested'],
+        ignore: ['after-comment'],
+        ignoreAtRules: ['value'],
+      },
+    ],
   },
   overrides: [
     {